add numpy fitting refactor
